move a couple helpers from Share.NamespaceDiffs to Share.NamespaceDiffs.Types
